RAG/requirements.txt
langgraph
groq
faiss-cpu
numpy
sentence-transformers
PyPDF2
fastapi 
uvicorn
duckduckgo-search 